requirements.txt
nextcord
squarecloud-api